Ran into an issue when finally looking at getting the wallet to whitelist rings as acceptable contents. As I looked at the whitelist... ID cards appeared to already be whitelisted. But in Dev environment... They were not. Had to figure out why with Dan, appears ID cards do not actually have the "IDCardStandard" component on them anywhere, so they never actually would fit.
Using in game component exam, we found that the ID Cards use actually a "IdCard" component instead. Whitelisting for that resolved the issue. Leaving IDCardStandard there incase it has another use for something that we haven't considered for. Now players should be able to fit their ID, or spare ID's in their wallet for safe keeping.