Locale tags.ftl is directly from @drraccoony. I had no idea even where to look for this. Thank you.
Had a LOT of struggle with this, but it's finally PR-able. A lot of oversighted portions got this backed up. Still learning, and pretty much all of my issues have been because I forgot to backtrace a component so that they point to each other properly, or having case sensitive issues. I'm comfortable putting this out in the state it's in now though. It works great, and will be perfect for the festival. I really hope people like this new little toy.

The only thing it may be missing, is I don't know how it will handle with space cleanup if people put these outside. They can walk on their own, and will burn for a long while before stopping.